requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
requests==2.32.3
twilio==9.3.6
